Fix Bug
The button corresponding to the current generation is now disabled. So if the user is on "Page 2", the user can't click "Page 2" again. This was important because currently we fetch data when "gen" state changes and if the user clicks on the same gen in which he/she is currently at, the loader gets activated but we don't fetch data because "gen" state remains the same and therefore the site is stuck with an infinite loading animation.